+++ src/probe_matcher.h
@@ -111,5 +111,6 @@ private:
       const std::set<std::string> &tracepoints);
 
   FuncParamLists get_iters_params(const std::set<std::string> &iters);
+  FuncParamLists get_rawtracepoint_params(const std::set<std::string> &raw_tps);
 };
 } // namespace bpftrace