+++ CHANGELOG.md
@@ -40,6 +40,8 @@ and this project adheres to
   - [#3863](https://github.com/bpftrace/bpftrace/pull/3863)
 - Add license config to specify BPF license
   - [#3905](https://github.com/bpftrace/bpftrace/pull/3905)
+- Rawtracepoints can now use `args` builtin and list params
+  - [#3918](https://github.com/bpftrace/bpftrace/pull/3918)
 #### Changed
 - `probe` builtin is now represented as a string type
   - [#3638](https://github.com/bpftrace/bpftrace/pull/3638)